+++ lib/flutter_crypto_security.dart
@@ -4,4 +4,5 @@ library;
 
 export 'crypto.dart';
 export 'src/aes_gcm.dart' show AesGcm;
+export 'src/envelope_v2.dart' show CryptoV2;
 export 'src/rsa_keys.dart' show RsaKeyCodec;